requirements.txt
agora-token-builder==1.0.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==3.2.16
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.1
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
mypy-extensions==0.4.3
pathspec==0.10.3
phonenumbers==8.13.2
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
pytz==2022.7
redis==4.4.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.3.0
